unsigned suffices for num SIMD lanes

+++ include/visionaray/get_color.h
@@ -123,7 +123,7 @@ inline vector<3, T> get_color(
     float_array y;
     float_array z;
 
-    for (size_t i = 0; i < simd::num_elements<T>::value; ++i)
+    for (unsigned i = 0; i < simd::num_elements<T>::value; ++i)
     {
         auto c = hrs[i].hit ? colors[hrs[i].prim_id] : C();
         x[i] = c.x;
@@ -181,7 +181,7 @@ inline auto get_color(
     float_array y3;
     float_array z3;
 
-    for (size_t i = 0; i < simd::num_elements<U>::value; ++i)
+    for (unsigned i = 0; i < simd::num_elements<U>::value; ++i)
     {
         auto cc1 = get_clr(i, 0);
         auto cc2 = get_clr(i, 1);
